refactor: Restructure project
The following changes have been made:
- all public interfaces are included in EnvPicker.kt
- Config is a simple data structure now, previously included code has been moved into classes that encapsulate it better
- Tests of EntryReflection and EntrySerializer have been extract to their own classes
