+++ env-picker/src/main/java/de/br/android/envpicker/EnvPicker.kt
@@ -1,15 +1,11 @@
 package de.br.android.envpicker
 
 import android.content.Context
-import android.content.Intent
 import androidx.fragment.app.Fragment
-import de.br.android.envpicker.ui.EnvActivity
 import de.br.android.envpicker.ui.EnvFragment
 
-
 /**
- * An [EnvPicker] can be used to manage its data directly or create a [Fragment] that
- * lets users manage the data via UI.
+ * An [EnvPicker] stores environment variables and provides a convenient management UI.
  */
 interface EnvPicker<T : Entry> {
     /**
@@ -52,8 +48,7 @@ interface EnvPicker<T : Entry> {
 }
 
 /**
- * Instantiates an [EnvPicker] that can be used to manage its data directly or create a [Fragment]
- * that lets users manage the data via UI.
+ * Creates an instance of [EnvPicker].
  *
  * @param config a valid [Config] object
  * @param context to be used to access SharedPreferences and resources
@@ -62,54 +57,10 @@ interface EnvPicker<T : Entry> {
  * @sample de.br.android.envpicker.envPickerSample
  */
 fun <T : Entry> envPicker(config: Config<T>, context: Context): EnvPicker<T> =
-    object : EnvPicker<T> {
-
-        init {
-            config.validate()
-            ConfigStore.set(config.key, config)
-            setupDefaultEntries(context)
-        }
-
-        @Suppress("UNCHECKED_CAST")
-        private fun getRepo(context: Context) =
-            (EnvRepository.get(config, context) as EnvRepository<T>)
-
-        override val config = config
-
-        override fun createFragment() = EnvFragment.create<T>(config.key)
-
-        override fun getEntries(context: Context): List<T> =
-            getRepo(context).loadEntries()
-
-        override fun setEntries(state: List<T>, context: Context) =
-            getRepo(context).saveEntries(state)
-
-        override fun getActiveEntry(context: Context): T =
-            getRepo(context).loadActiveEntry()
-
-        override fun setActiveEntry(entry: T, context: Context) {
-            getRepo(context).saveActiveEntry(entry)
-        }
-
-        override fun startEnvPickerActivity(context: Context) {
-            context.startActivity(Intent(context, EnvActivity::class.java).apply {
-                putExtra(ConfigStore.KEY, config.key)
-            })
-        }
-
-        private fun setupDefaultEntries(context: Context) {
-            if (getEntries(context).isNullOrEmpty()) {
-                setEntries(config.defaultEntries, context)
-                setActiveEntry(
-                    config.defaultActiveEntry, context
-                )
-            }
-        }
-    }
+    EnvPickerImpl(config, context)
 
 /**
- * Instantiates an [EnvPicker] that can be used to manage its data directly or create a [Fragment]
- * that lets users manage the data via UI.
+ * Creates an instance of [EnvPicker].
  *
  * @property key Used as SharedPreferences key. Needs to be unique among [EnvPicker] instances.
  * @property fragmentTitle Displayed as title of fragment.
@@ -139,3 +90,69 @@ fun <T : Entry> envPicker(
         ),
         context,
     )
+
+/**
+ * Holds the complete configuration of an EnvPicker instance.
+ *
+ * @property key Used as SharedPreferences key. Needs to be unique among [EnvPicker] instances.
+ * @property fragmentTitle Displayed as title of fragment.
+ * @property defaultEntries Will be set as initial values if no entries are present at init time.
+ * @property defaultActiveEntry Will be set as initial active entry.
+ * @property customSerializer Will be used to handle serialization of the entries.
+ * @property clearOnChangedDataFormat If true, clears the persisted entries in case of a format change.
+ */
+data class Config<T : Entry>(
+    val key: String,
+    val fragmentTitle: String,
+    val defaultEntries: List<T> = listOf(),
+    val defaultActiveEntry: T,
+    val customSerializer: EntrySerializer<T>?,
+    val clearOnChangedDataFormat: Boolean,
+)
+
+/**
+ * Needs to be applied to all parameters of an [Entry]'s primary constructor.
+ * @param label will appear in the UI as label for the corresponding field.
+ */
+@Target(AnnotationTarget.VALUE_PARAMETER)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class EntryField(val label: String)
+
+/**
+ * An entry to be managed and persisted by the EnvPicker library.
+ */
+interface Entry {
+    /**
+     * The name that will be displayed in the entry selection screen.
+     */
+    val name: String
+
+    /**
+     * The description that will be displayed in the entry selection screen.
+     */
+    val summary: String get() = EntryReflection.getDefaultSummary(this)
+}
+
+/**
+ * An implementation of [Entry] representing a single key-value pair.
+ *
+ * @param name The unique name of this [Entry] displayed in the UI and used as key
+ * @param value The [String] value of this [Entry]
+ */
+data class KeyValueEntry(
+    @EntryField("Name")
+    override var name: String,
+    @EntryField("Value")
+    var value: String,
+) : Entry
+
+/**
+ * Defines methods to serialize an implementation of Entry into a String and vice versa.
+ *
+ * @property serializeEntry used to serialize the [Entry] implementation
+ * @property deserializeEntry used to deserialize the [Entry] implementation
+ */
+interface EntrySerializer<T : Entry> {
+    fun serializeEntry(entry: T): String
+    fun deserializeEntry(str: String): T
+}
